@@ -1,6 +1,6 @@
 require 'test_helper'
 
-class LoginsControllerTest < ActionController::TestCase
+class LoginsControllerTest < ActionDispatch::IntegrationTest
   setup do
     @login = logins(:login_one)
     login_as :login_one
@@ -8,20 +8,20 @@ class LoginsControllerTest < ActionController::TestCase
 
   test 'should create login' do
     assert_difference('Login.count') do
-      post :create, login: { password: 'secret', username: 'asdf' }
+      post logins_url, params: { login: { password: 'secret', username: 'asdf' } }
     end
 
     assert_redirected_to posts_path
   end
 
   test 'should update login' do
-    patch :update, id: @login, login: { password: 'secretss', username: 'newusername' }
+    patch login_url(@login), params: { login: { password: 'secretss', username: 'newusername' } }
     assert_redirected_to login_path(@login)
   end
 
   test 'should destroy login' do
     assert_difference('Login.count', -1) do
-      delete :destroy, id: @login
+      delete login_url(@login)
     end
 
     assert_redirected_to logins_path
