@@ -1,34 +1,36 @@
 require 'test_helper'
 
-class PostsControllerTest < ActionController::TestCase
+class PostsControllerTest < ActionDispatch::IntegrationTest
   setup do
     @post = posts(:post_one)
     login_as :login_one
   end
 
   test 'should get index' do
-    get :index
+    get posts_url
     assert_response :success
     assert_not_nil assigns(:posts)
   end
 
   test 'should get archive' do
-    get :archive
+    get archive_url
     assert_response :success
     assert_not_nil assigns(:posts)
   end
 
   test 'should get new' do
-    get :new
+    get new_post_url
     assert_response :success
   end
 
   test 'should create post' do
     assert_difference('Post.count') do
-      post :create, post: {
-        contents: @post.contents,
-        title: @post.title,
-        slug: 'new-slug'
+      post posts_url, params: {
+        post: {
+          contents: @post.contents,
+          title: @post.title,
+          slug: 'new-slug'
+        }
       }
     end
 
@@ -36,27 +38,29 @@ class PostsControllerTest < ActionController::TestCase
   end
 
   test 'should show post' do
-    get :show, id: @post
+    get post_url(@post)
     assert_response :success
   end
 
   test 'should get edit' do
-    get :edit, id: @post
+    get edit_post_url(@post)
     assert_response :success
   end
 
   test 'should update post' do
-    patch :update, id: @post, post: {
-      contents: @post.contents,
-      title: @post.title,
-      slug: @post.slug
+    patch post_url(@post), params: {
+      post: {
+        contents: @post.contents,
+        title: @post.title,
+        slug: @post.slug
+      }
     }
     assert_redirected_to post_path(assigns(:post))
   end
 
   test 'should destroy post' do
     assert_difference('Post.count', -1) do
-      delete :destroy, id: @post
+      delete post_url(@post)
     end
 
     assert_redirected_to posts_path
